feat: add metadata
- fix trivial problems
