+++ app/profile/nickname/page.tsx
@@ -21,7 +21,7 @@ const NicknamePage = () => {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
-                'Authorization': token
+                'Authorization': `${token}`
             },
         });
 
@@ -37,7 +37,7 @@ const NicknamePage = () => {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
-                'Authorization': token
+                'Authorization': `${token}`
             },
             body : JSON.stringify({
                 nickName : nickname
