+++ app/profile/page.tsx
@@ -21,7 +21,7 @@ const Page = () => {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json',
-                'Authorization': token
+                'Authorization': `${token}`
             },
         });
